+++ RateLimited/Core/Time/Clock.swift
@@ -5,6 +5,9 @@ protocol Clock: Sendable {
 }
 
 struct SystemClock: Clock {
+    // swiftlint:disable:next unneeded_synthesized_initializer
+    nonisolated init() {}
+
     nonisolated func now() -> Date {
         Date()
     }